pyproject.toml
[tool.poetry]
name = "asana-archie"
version = "0.1.0"
description = "A framework for automating Asana projects and workflows."
authors = ["Joe Trollo <joetrollo@asana.com>"]
license = "MIT"

readme = "README.rst"

homepage = "https://gitlab.com/joetrollo/archie"
repository = "https://gitlab.com/joetrollo/archie"
documentation = "https://joetrollo.gitlab.io/archie"

packages = [
    { include = "archie" },
]

[tool.poetry.dependencies]
python = "^3.7"
cattrs = "^0.9"
asana = "^0.9"
attrs = "^19.1"
typing_extensions = "^3.7"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allow-prereleases = true}
coverage = "^4.5"
freezegun = "^0.3.12"
isort = "^4.3"
mypy = "^0.720.0"
pyflakes = "^2.1"
radon = "^2.4"
sphinx = "^2.2"
sphinx-autodoc-typehints = "^1.8"
sphinx-rtd-theme = "^0.4.3"
twine = "^2.0"
xenon = "^0.5.5"

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
